renamed files; added header documentation
